update GPAName source field
Identified submission requests where controlled access is true and GPAName is empty, null, or missing. For these submission requiests it iterates through questionarieData.study.funding array and uses the first non-empty nciGPA field found to initialize the root GPAName field. If that is not found then questionarieData.study.GPAName is checked as a fallback (expected that this scenario does not exist).